Add exceptions handlers

+++ src/main/java/com/devsuperior/DScommerce/services/ProductService.java
@@ -2,14 +2,20 @@ package com.devsuperior.DScommerce.services;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.DScommerce.domain.Product;
 import com.devsuperior.DScommerce.dto.ProductDTO;
 import com.devsuperior.DScommerce.repositories.ProductRepository;
+import com.devsuperior.DScommerce.services.exceptions.DatabaseException;
+import com.devsuperior.DScommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -19,7 +25,8 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = productRepository.findById(id).get();
+		Product product = productRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(product);
 	}
 
@@ -41,18 +48,31 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = productRepository.getReferenceById(id);
+		try {
+			Product entity = productRepository.getReferenceById(id);
 
-		entity.setName(dto.getName());
-		entity.setDescription(dto.getDescription());
-		entity.setPrice(dto.getPrice());
-		entity.setImgUrl(dto.getImgUrl());
-		return new ProductDTO(entity);
+			entity.setName(dto.getName());
+			entity.setDescription(dto.getDescription());
+			entity.setPrice(dto.getPrice());
+			entity.setImgUrl(dto.getImgUrl());
+			return new ProductDTO(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		productRepository.deleteById(id);
+		if(!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		
+		try {
+			productRepository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 
 	}
 }
